The call to fabs(long double) is ambiguous in Ubuntu 64 bits. Replaced with fabs(double).

+++ v3d_main/basic_c_fun/vcdiff.cpp
@@ -50,7 +50,7 @@ int fabs(int x)
 
 V3DLONG fabs(V3DLONG x)
 {
-	return (V3DLONG)(fabs((long double)x));
+	return (V3DLONG)(fabs((double)x));
 }
 
 double log(int x)
